requirements.txt
paho-mqtt
aiohttp
shortuuid
